+++ tide/core/ContentWindowController.h
@@ -51,6 +51,7 @@
 enum SizeState
 {
     SIZE_1TO1,
+    SIZE_1TO1_FITTING,
     SIZE_LARGE,
     SIZE_FULLSCREEN
 };